Get file watcher .add() to work on Linux
